fixes on scripts, geki story improvements, code optimizations

+++ scripts/cleanup.sh
@@ -1,16 +1,22 @@
 #!/bin/bash
 
-cd /var/srvcs/services/definitions
+FOLDER = /var/srvcs/services/definitions
 
-for filename in *.service; do
-	noext=${filename%%.*}
-	sudo systemctl stop $filename
-	sudo systemctl disable $filename
-	sudo rm -f /lib/systemd/system/$filename
-done
+if [ -d $FOLDER ]
+then
+    cd $FOLDER
 
-cd ../
-sudo rm -rf definitions/*
-sudo rm -rf scripts/*
+	for filename in *.service; do
+		noext=${filename%%.*}
+		sudo systemctl stop $filename
+		sudo systemctl disable $filename
+		sudo rm -f /lib/systemd/system/$filename
+	done
+
+	cd ../
+	sudo rm -rf definitions/*
+	sudo rm -rf scripts/*
+
+fi
 
 # sudo systemctl daemon-reload
